feat(cli): add `tauri bundle` subcommand
closes #8734
