+++ .changes/tauri-bundle-command.md
@@ -1,0 +1,7 @@
+---
+"tauri-cli": "patch:feat"
+"@tauri-apps/cli": "patch:feat"
+---
+
+Add `tauri bundle` subcommand which runs the bundle phase only, best paired with `tauri build --no-bundle`
+
